+++ webapp/src/main/java/ar/edu/itba/paw/webapp/form/CreatePromotionForm.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 @EndDateTimeAfterStartDateTime(startDateTimeField = "startDateTime", endDateTimeField = "endDateTime")
 public class CreatePromotionForm {
     @NotNull
-    private Integer sourceProductId;
+    private Long sourceProductId;
 
     @NotNull
     @Min(1)
@@ -28,11 +28,11 @@ public class CreatePromotionForm {
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
     private LocalDateTime endDateTime;
 
-    public Integer getSourceProductId() {
+    public Long getSourceProductId() {
         return sourceProductId;
     }
 
-    public void setSourceProductId(Integer sourceProductId) {
+    public void setSourceProductId(long sourceProductId) {
         this.sourceProductId = sourceProductId;
     }
