+++ webapp/src/main/java/ar/edu/itba/paw/webapp/form/EditCategoryOrderForm.java
@@ -5,7 +5,7 @@ import javax.validation.constraints.NotNull;
 
 public class EditCategoryOrderForm {
     @NotNull
-    private Integer restaurantId;
+    private Long restaurantId;
 
     @NotNull
     @Min(0)
@@ -15,11 +15,11 @@ public class EditCategoryOrderForm {
     @Min(0)
     private Integer newOrder;
 
-    public Integer getRestaurantId() {
+    public Long getRestaurantId() {
         return restaurantId;
     }
 
-    public void setRestaurantId(Integer restaurantId) {
+    public void setRestaurantId(Long restaurantId) {
         this.restaurantId = restaurantId;
     }
